Able to obtain total from two sample receipts
