+++ scanner/tests/test_core.py
@@ -5,5 +5,5 @@ from unittest import TestCase
 class TestCore(TestCase):
 
     def test_ocrProcessing(self):
-        self.assertEqual("A Python Approach to Character\nRecognition",
-                         core.ocrProcessing('scanner\\tests\\resources\\test_image.png'))
+        self.assertEqual(['46.13'], core.ocrProcessing('scanner\\tests\\resources\\receipt_2.jpg'))
+        self.assertEqual(['81.14'], core.ocrProcessing('scanner\\tests\\resources\\receipt_3.jpg'))
